accounts working

+++ code/SQL/db_build.sql
@@ -18,7 +18,7 @@ FOREIGN KEY (exchange_name) REFERENCES Exchange(name)
 CREATE TABLE Account(
 user_id INT,
 exchange_name CHAR(20),
-value INT,
+value DOUBLE,
 PRIMARY KEY (user_id, exchange_name),
 FOREIGN KEY (user_id) REFERENCES User(id),
 FOREIGN KEY (exchange_name) REFERENCES Exchange(name)
